add factory.py for functions to create components automatically
