+++ limbus/core/__init__.py
@@ -1,4 +1,5 @@
-from .core import ComponentsManager, Component, DefaultParam, ComponentState, Params, component_factory, NoValue
+from .core import ComponentsManager, Component, DefaultParam, ComponentState, Params, NoValue
+from .factory import component_factory
 
 __all__ = [
     "ComponentsManager",
